Lobby screen buttons
Working buttons switching between the different screens within the lobby screen. This is massive cheese method I have done but it works so it works.

There is some bugs with it but only really when the user is spamming faster than the function can be called.

This is sort of unavoidable and, well not my issue.