Time spent on importing
Quests, Rhu-Pass and Locker elements have been imported. It is now for bits of code and finish up these screens with code but before the next push will be getting the rest of the screens in. I don't think this will be done by the formative but I may try